Update dependencies and enhance UI in LessonLevel.vue
Updated all the dependencies versions in package.json to their latest stable versions for improved performance and security compliance.

Added a new UI layout in LessonLevel.vue to provide a better visual representation of lessons. Now, lessons are represented as a list of cards instead of a dropdown menu for a more user-friendly and intuitive interface.

Auto-login functionality added in user.js store after registration. After a user registers for a new account, the application will automatically log them in, providing a better user experience.

This commit also includes updates on the generated yarn.lock file.